pretty logging.
git-svn-id: http://www.virtualbox.org/svn/vbox/trunk@14958 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/recompiler_new/VBoxRecompiler.c
@@ -3178,7 +3178,7 @@ int32_t remR3PhysReadS16(RTGCPHYS SrcGCPhys)
     val = PGMR3PhysReadU16(cpu_single_env->pVM, SrcGCPhys);
     STAM_PROFILE_ADV_STOP(&gStatMemRead, a);
 #ifdef VBOX_DEBUG_PHYS
-    LogRel(("reads16: %x <- %08x\n", val, (uint32_t)SrcGCPhys));
+    LogRel(("reads16: %x <- %08x\n", (uint16_t)val, (uint32_t)SrcGCPhys));
 #endif
     return val;
 }
